backend/requirements.txt
fastapi
uvicorn
pydantic
alembic
psycopg2-binary
python-jose[cryptography]
python-multipart
python-dotenv
pydantic-settings
email-validator
passlib==1.7.4
bcrypt==3.2.2
sqlalchemy
